updated settings to Xcode 6.1 (6A1052d) renamed 'description' UITextView to 'details' (prevents a crash) eliminated compiler warnings

+++ MagazineView/Magazine Collection View/MagazineLayout.m
@@ -100,7 +100,7 @@
 {
     CGRect frame = [self frameForLayoutType:[self layoutTypeForIndexPath:indexPath]];
     
-    int row = indexPath.row / MagazineLayoutsPerRow;
+    long row = indexPath.row / MagazineLayoutsPerRow;
     frame.origin.y += row * [self rowHeight];
     
     UICollectionViewLayoutAttributes* attributes = [UICollectionViewLayoutAttributes layoutAttributesForCellWithIndexPath:indexPath];
